@@ -23,52 +23,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.1",
-    "@angular/common": "^5.0.1",
-    "@angular/compiler": "^5.0.1",
-    "@angular/core": "^5.0.1",
-    "@angular/forms": "^5.0.1",
-    "@angular/http": "^5.0.1",
-    "@angular/platform-browser": "^5.0.1",
-    "@angular/platform-browser-dynamic": "^5.0.1",
-    "@angular/router": "^5.0.1",
+    "@angular/animations": "6.1.10",
+    "@angular/common": "6.1.10",
+    "@angular/core": "6.1.10",
+    "@angular/forms": "6.1.10",
+    "@angular/http": "6.1.10",
+    "@angular/platform-browser": "6.1.10",
+    "@angular/platform-browser-dynamic": "6.1.10",
+    "@angular/router": "6.1.10",
     "angular2-image-zoom": "^1.2.1",
     "autocomplete.js": "^0.29.0",
     "bulma": "^0.6.1",
     "bulma-badge": "0.0.3",
-    "core-js": "^2.4.1",
-    "handlebars": "^4.0.11",
-    "mdi": "^2.0.46",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "core-js": "^2.5.7",
+    "handlebars": "^4.0.12",
+    "mdi": "^2.2.43",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.2",
-    "@angular/compiler-cli": "^5.0.1",
-    "@angular/language-service": "^5.0.1",
+    "@angular-devkit/build-angular": "^0.8.5",
+    "@angular/cli": "6.2.5",
+    "@angular/compiler": "6.1.10",
+    "@angular/compiler-cli": "6.1.10",
+    "@angular/language-service": "6.1.10",
     "@cxcloud/ct-types": "^0.1.6",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "@types/jasmine": "^2.8.9",
+    "@types/jasminewd2": "^2.0.5",
+    "@types/node": "^10.12.0",
+    "codelyzer": "^4.5.0",
     "husky": "^0.14.3",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "jasmine-core": "^3.2.1",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^3.0.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "^1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.3.1",
     "lint-staged": "^4.3.0",
     "mkdirp": "^0.5.1",
-    "ngx-markdown": "^1.4.0",
+    "ngx-markdown": "^1.6.0",
     "ngx-webstorage": "^2.0.0",
-    "prettier": "^1.8.2",
-    "protractor": "~5.1.2",
-    "serve": "^6.4.1",
+    "prettier": "^1.14.3",
+    "protractor": "^5.4.1",
+    "rxjs-tslint": "^0.1.5",
+    "serve": "^6.5.8",
     "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "^2.4.2"
+    "tslint": "^5.11.0",
+    "typescript": "2.9.2"
   }
 }
