@@ -1,23 +1,22 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { FormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { Ng2Webstorage } from 'ngx-webstorage';
-
-import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
 import { CartModule } from './cart';
-import { CoreModule } from './core';
 import { CategoriesModule } from './categories';
+import { CheckoutModule } from './checkout';
+import { CoreModule } from './core';
 import { HomeModule } from './home';
 import { LayoutModule } from './layout';
-import { SharedModule } from './shared';
-import { CheckoutModule } from './checkout';
 import { LoginModule } from './login';
-import { PagesModule } from './pages';
 import { OrderModule } from './order';
+import { PagesModule } from './pages';
 import { SearchModule } from './search';
+import { SharedModule } from './shared';
 
 @NgModule({
   imports: [
