@@ -1,6 +1,6 @@
-import { CartComponent } from './cart.component';
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
+import { CartComponent } from './cart.component';
 
 const routes: Routes = [{ path: 'cart', component: CartComponent }];
 
