@@ -1,11 +1,10 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-import { CategoriesRoutingModule } from './categories-routing.module';
-import { ProductListComponent } from './product-list';
-import { ProductComponent } from './product';
-import { SharedModule } from '../shared';
+import { NgModule } from '@angular/core';
 import { ImageZoomModule } from 'angular2-image-zoom';
+import { SharedModule } from '../shared';
+import { CategoriesRoutingModule } from './categories-routing.module';
+import { ProductComponent } from './product';
+import { ProductListComponent } from './product-list';
 
 const DECLARATIONS = [ProductListComponent, ProductComponent];
 
