@@ -1,3 +1,3 @@
+export * from './categories.module';
 export * from './product';
 export * from './product-list';
-export * from './categories.module';
