@@ -1,13 +1,12 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
 import { SharedModule } from '../shared';
 import { CheckoutRoutingModule } from './checkout-routing.module';
 import { CheckoutComponent } from './checkout.component';
+import { ConfirmationComponent } from './confirmation';
 import { ShippingComponent } from './shipping';
 import { PaymentComponent } from './shipping/payment';
-import { ConfirmationComponent } from './confirmation';
 import { ShippingInfoComponent } from './shipping/shipping-info';
 
 const DECLARATIONS = [
