@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Cart, LineItem } from '@cxcloud/ct-types/carts';
+import { Cart } from '@cxcloud/ct-types/carts';
 import { Address } from '@cxcloud/ct-types/common';
 import { CartService } from '../../core/cart/cart.service';
 import { OrderService } from '../../core/order/order.service';
