@@ -1,3 +1,3 @@
+export * from './checkout.module';
 export * from './confirmation';
 export * from './shipping';
-export * from './checkout.module';
