@@ -1,13 +1,10 @@
-import { Injectable } from '@angular/core';
-import { Headers } from '@angular/http';
 import { HttpClient } from '@angular/common/http';
-
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-
+import { Injectable } from '@angular/core';
 import { Category } from '@cxcloud/ct-types/categories';
 import { PaginatedProductResult, Product } from '@cxcloud/ct-types/products';
 import { ShippingMethod } from '@cxcloud/ct-types/shipping';
+import { Observable } from 'rxjs';
+import { tap } from 'rxjs/operators';
 import { sortByCustomValues } from '../../utils/helpers';
 
 @Injectable()
@@ -17,7 +14,9 @@ export class CommerceService {
   getCategories(): Observable<Category[]> {
     return this.http
       .get<Category[]>('/categories')
-      .do(categories => this.sortCategories(categories));
+      .pipe(
+        tap(categories => this.sortCategories(categories))
+      );
   }
 
   getProducts(categoryId: string): Observable<PaginatedProductResult> {
