@@ -1,15 +1,13 @@
 import { CommonModule } from '@angular/common';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
-
 import { AuthInterceptor } from './auth/auth.interceptor';
-import { CurrentUserService } from './auth/current-user.service';
 import { AuthService } from './auth/auth.service';
+import { CurrentUserService } from './auth/current-user.service';
 import { CartService } from './cart/cart.service';
-import { OrderService } from './order/order.service';
 import { CommerceService } from './commerce/commerce.service';
 import { ContentService } from './content/content.service';
+import { OrderService } from './order/order.service';
 import { SearchService } from './search/search.service';
 
 @NgModule({
