@@ -1,6 +1,5 @@
-import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Injectable } from '@angular/core';
 import { environment } from '../../../environments/environment';
 import { buildUrlQuery } from '../../utils/helpers';
 
