@@ -1,10 +1,9 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-
-import { HomeComponent } from './home.component';
-import { HomeRoutingModule } from './home-routing.module';
 import { SharedModule } from '../shared';
+import { HomeRoutingModule } from './home-routing.module';
+import { HomeComponent } from './home.component';
 
 @NgModule({
   imports: [CommonModule, FormsModule, SharedModule, HomeRoutingModule],
