@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { CartService } from '../../core/cart/cart.service';
-import { CurrentUserService } from '../../core/auth/current-user.service';
 import { AuthService } from '../../core/auth/auth.service';
+import { CurrentUserService } from '../../core/auth/current-user.service';
+import { CartService } from '../../core/cart/cart.service';
 
 @Component({
   selector: 'app-brand-navbar',
