@@ -1,5 +1,6 @@
 export * from './brand-navbar';
 export * from './footer';
 export * from './header';
-export * from './main-menu';
 export * from './layout.module';
+export * from './main-menu';
+
