@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { MainMenuComponent } from './main-menu.component';
 import { RouterTestingModule } from '@angular/router/testing';
+import { MainMenuComponent } from './main-menu.component';
 
 describe('MainMenuComponent', () => {
   let component: MainMenuComponent;
