@@ -1,10 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Order } from '@cxcloud/ct-types/orders';
 import { Address } from '@cxcloud/ct-types/common';
+import { Order } from '@cxcloud/ct-types/orders';
 import { CartService } from '../core/cart/cart.service';
 import { OrderService } from '../core/order/order.service';
-import { DatePipe } from '@angular/common';
 
 @Component({
   selector: 'app-order',
