@@ -1,7 +1,6 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
 import { SharedModule } from '../shared';
 import { OrderRoutingModule } from './order-routing.module';
 import { OrderComponent } from './order.component';
