@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
 import { PageComponent } from './page/page.component';
 
 const routes: Routes = [{ path: 'pages/:pageId', component: PageComponent }];
