@@ -1,5 +1,4 @@
-import { Component, OnInit, Input } from '@angular/core';
-import 'rxjs/add/operator/debounceTime';
+import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-error-display',
