@@ -1,21 +1,20 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CheckoutButtonComponent } from './checkout-button.component';
+import { GotoButtonComponent } from './goto-button.component';
 
 describe('CheckoutButtonComponent', () => {
-  let component: CheckoutButtonComponent;
-  let fixture: ComponentFixture<CheckoutButtonComponent>;
+  let component: GotoButtonComponent;
+  let fixture: ComponentFixture<GotoButtonComponent>;
 
   beforeEach(
     async(() => {
       TestBed.configureTestingModule({
-        declarations: [CheckoutButtonComponent]
+        declarations: [GotoButtonComponent]
       }).compileComponents();
     })
   );
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CheckoutButtonComponent);
+    fixture = TestBed.createComponent(GotoButtonComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
