@@ -4,7 +4,7 @@ export * from './goto-button';
 export * from './info-card';
 export * from './order-details';
 export * from './quantity-input';
-export * from './spinner';
-export * from './steps';
-export * from './stars';
 export * from './shared.module';
+export * from './spinner';
+export * from './stars';
+export * from './steps';
