@@ -9,7 +9,5 @@ export class SpinnerComponent implements OnInit {
 
   constructor() {}
 
-  ngOnInit() {
-  }
-
+  ngOnInit() {}
 }
