@@ -1,5 +1,4 @@
-import { any } from 'codelyzer/util/function';
-import { Component, OnInit,Input } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-steps',
